We may use SELECTED date dynamically now.
Before we manually put the date "Sunday" to numbering how many Sundays are there between the start and end date. So, now we may select any of the dates from our front end and don't need to change over the coursecreate.php file anymore. 

Actually, format("l") provides an example: Sunday but we stored our data in small-case and that's why need to use ucfirst($day) to equivalent our data. So, it works. 

Thanks
Roman